feat: WHY IS THERE CODE???
I DONT GIVE A FUCK ABOUT THE FUCKING CODE!
i just want to download this stupid fucking application and use it
https://github.com/sherlock-project/sherlock#installation

WHY IS THERE CODE??? MAKE A FUCKING .EXE FILE AND GIVE IT TO ME.
these dumbfucks think that everyone is a developer and understands code.
well i am not and i don't understand it. I only know to download and
install applications. SO WHY THE FUCK IS THERE CODE?
make an EXE file and give it to me. STUPID FUCKING SMELLY NERDS